pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lcmeval"
version = "0.1.0"
description = "LCM Evaluation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "numpy==2.2.5",
    "openai>=1.76.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[tool.setuptools]
packages = [
    "lcmeval",
    "lcmeval.crawler",
    "lcmeval.utils",
]
